feat(update-jug): move files
