+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { BrowserRouter } from 'react-router-dom'
 import { StylesProvider } from '@material-ui/styles'
 import Juggernaut from 'containers/Juggernaut'
-import theme from 'commons/constants/theme'
+import theme from 'common/constants/theme'
 import Main from 'containers/Main'
 
 const App = () => (
