+++ src/assets/images/tabs.png
