+++ src/components/Button/index.jsx
@@ -1,0 +1,3 @@
+import Button from '@material-ui/core/Button'
+
+export default Button
