--- src/components/Loading/__tests__/LoadingStyled.test.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { mount } from 'enzyme'
-import LoadingStyled from '../styled'
-
-describe('Loading styled', () => {
-  it('should render with color', () => {
-    const component = mount(<LoadingStyled />)
-
-    expect(component).toHaveStyleRule('color', '#777')
-  })
-})
-
