--- src/components/Loading/style.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-const LoadingStyled = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`
-
-export default LoadingStyled
