+++ src/components/Loading/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+  align-items: center;
+  display: flex;
+  height: 100vh;
+  justify-content: center;
+  width: 100%;
+`
