+++ src/containers/Home/index.jsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import { Grid, Typography, CardActions, Button, Container } from '@material-ui/core'
-import signIn from 'assets/img/sign-in.png'
-import signInSide from 'assets/img/sign-in-side.png'
-import signUp from 'assets/img/sign-up.png'
-import modal from 'assets/img/modal.png'
-import menu from 'assets/img/menu.png'
-import table from 'assets/img/table.png'
-import card from 'assets/img/card.png'
-import tabs from 'assets/img/tabs.png'
-import toolbar from 'assets/img/toolbar.png'
-import list from 'assets/img/list.png'
-import logo from 'assets/img/juggernaut.gif'
+import signIn from 'assets/images/sign-in.png'
+import signInSide from 'assets/images/sign-in-side.png'
+import signUp from 'assets/images/sign-up.png'
+import modal from 'assets/images/modal.png'
+import menu from 'assets/images/menu.png'
+import table from 'assets/images/table.png'
+import card from 'assets/images/card.png'
+import tabs from 'assets/images/tabs.png'
+import toolbar from 'assets/images/toolbar.png'
+import list from 'assets/images/list.png'
+import logo from 'assets/images/juggernaut.gif'
 import Copyright from './Copyright'
 import { GridItemStyled, CardStyled, CardMediaStyled, CardContentStyled, ImgStyled, HomeStyled } from './style'
