+++ src/resources/api/example-api.js
