+++ src/resources/example.js
